% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{.gsea_barplot_down}
\alias{.gsea_barplot_down}
\title{Plot barplot to plot GSEA barplot down}
\usage{
.gsea_barplot_down(gseaRes, n = 15, fill = NA)
}
\description{
@param gseaRes GSEA results from run_gsea
@param n number of pathways to plot
@param fill fill color of bars
@noRd
}
