% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{gsea_barplot}
\alias{gsea_barplot}
\title{Wrapper function around various barplots for gsea}
\usage{
gsea_barplot(
  gseaRes,
  n = 15,
  direction = c("all", "up", "down"),
  fill = NA,
  return.type = c("patchwork", "list")
)
}
\arguments{
\item{gseaRes}{GSEA results from run_gsea}

\item{n}{number of pathways to plot}

\item{direction}{which direction to plot. all plots all,
both plots up and down in one graph
up and down plot only up and down respectively}

\item{fill}{what color to use for fill. Can be a vector of length 2,
a color string which will be reused or a column in GSEA res}

\item{return.type}{whether to return a patchwork object or a list of plots}

\item{sort.by}{what to sort by in the case of the both option}
}
\description{
Wrapper function around various barplots for gsea
}
