% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ortho_map}
\alias{ortho_map}
\title{Pull mouse equivalents of human gene sets}
\usage{
ortho_map(
  geneset,
  mart,
  keytype = "ensembl_gene_id",
  value = "mmusculus_homolog_ensembl_gene"
)
}
\arguments{
\item{geneset}{A list of genes to orthology map}

\item{mart}{a biomart object}

\item{keytype}{keytypes to search mart}

\item{value}{value to return}
}
\description{
Orthology map mouse to human gene sets using biomart
}
