% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{run_gsea}
\alias{run_gsea}
\title{Plot GSEA results}
\usage{
run_gsea(
  results,
  geneset,
  p = "PValue",
  logFC = "logFC",
  name = "gene_id",
  min.size = 10,
  max.size = 500,
  pcutoff = 0.05,
  eps = 1e-10
)
}
\arguments{
\item{results}{dataframe of gene expression}

\item{geneset}{set of genes to test, in the format GSNAME,GENES}

\item{p}{column with adjusted p values, default PValue}

\item{logFC}{column with logFC values}

\item{name}{column with gene name}

\item{min.size}{min gs size for GSEA}

\item{max.size}{max gs size for gsea}

\item{pcutoff}{P value cutoff for returned genes}

\item{eps}{boundary for calculating p value}

\item{ranking}{ranking metric for genes. "signp" for signlogFC*Pval, "FC" for logFC, logpfc for -logp\*FC}
}
\description{
Plot GSEA results
}
