% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_version}
\alias{strip_version}
\title{Strip version from ensenmbl geneID}
\usage{
strip_version(genes)
}
\arguments{
\item{genes}{Vector of ensembl geneid versions}
}
\description{
Strip version from ensenmbl geneID
}
